ジャケット画像の読み込みなどを実装

+++ Assets/Lyrid/Scripts/GameScene/Charts/Chart.cs
@@ -104,7 +104,7 @@ namespace Lyrid.GameScene.Charts
                             index++;
                             break;
                         case "#OFFSET":
-                            offsetTime = int.Parse(csvData[index][1]);
+                            offsetTime = float.Parse(csvData[index][1]);
                             index++;
                             break;
                         case "#BPM":
